+++ imports/i18n/locales/zh-CN.js
@@ -23,6 +23,7 @@ export const locale_zh_cn = {
   "RE": "回复",
   "Send Image": "图片",
   "Send Images": "多张图片",
+  "Send Markdown": "Markdown",
   "Send Image Message": "发送图片",
   "Image Message": "[图片]",
   "No content": "没有内容",